% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beast.R
\name{.read.beast}
\alias{.read.beast}
\title{read.beast}
\usage{
.read.beast(file)
}
\arguments{
\item{file}{beast file}
}
\value{
treedata object
}
\description{
read beast/mrbayes/mega Nexus output
}
\examples{
file <- system.file("extdata/BEAST", "beast_mcc.tree", package="treeio")
read.beast(file)
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
