% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\name{cluster_muts}
\alias{cluster_muts}
\title{Computes mutations for sequences within specified clusters}
\usage{
cluster_muts(
  Y,
  scanner_env,
  nodes = NULL,
  mutdata = list.files(patt = "cog_global_[0-9\\\\-]+_mutations.csv", full.name = TRUE,
    path = "../phylolatest/metadata/"),
  mut_variable = c("variants", "mutations"),
  min_seq_contrast = 1,
  overlap_threshold = 0.9,
  ncpu = 1
)
}
\arguments{
\item{Y}{summary table output of scanner}

\item{scanner_env}{output of scanner}

\item{nodes}{integer vector of nodes for which results will be computed. Computes for all nodes in Y if omitted}

\item{mutdata}{path to file or data frame with variant information}

\item{mut_variable}{Name of variable in data frame (spec by mutfn) which contains genetic variant data}

\item{ncpu}{number of CPUs for parallel processing}
}
\description{
Also derives defining mutations by contrasting with closest ancestor
}
