% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner_plots.R
\name{cluster_muts_new}
\alias{cluster_muts_new}
\title{mutation_plots}
\usage{
cluster_muts_new(
  scanner_env,
  nodes,
  mutfn = list.files(paste0(path_to_data), patt =
    "cog_global_[0-9\\\\-]+_mutations.csv", full.name = TRUE),
  min_seq_contrast = 1,
  overlap_threshold = 0.9
)
}
\description{
mutation_plots
}
