% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{cluster_origin_comparison}
\alias{cluster_origin_comparison}
\title{Compute cumulative proportion of samples descended from clusters originating within a specified time window}
\usage{
cluster_origin_comparison(
  s,
  mint = decimal_date(as.Date("2020-08-01")),
  maxt = Inf,
  res = 50
)
}
\arguments{
\item{s}{data frame with  del_introduction(character), sample_time(numeric)}

\item{mint}{lower sample time bound. Clusters must have sample before this time}

\item{maxt}{upper sample time bound. Clusters must have sample after this time}

\item{res}{integer resolution of time axis}
}
\value{
list with 1) result with  columns: cumulative fraction of samples descended from new clusters, number descended from old clusters, number descended from new clusters 2) data with cluster subset
}
\description{
Compute cumulative proportion of samples descended from clusters originating within a specified time window
}
