% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{cluster_origin_comparison2}
\alias{cluster_origin_comparison2}
\title{Compute cumulative proportion of samples descended from clusters originating within a specified time window or uk lineage}
\usage{
cluster_origin_comparison2(
  s,
  uk_lineage = NULL,
  genotype = NULL,
  mint = decimal_date(as.Date("2020-07-01")),
  maxt = Inf,
  detailed = TRUE
)
}
\arguments{
\item{s}{data frame with  del_introduction(character), sample_time(numeric), and epi_week (integer)}

\item{uk_lineage}{optional character vector of uk_lineage}

\item{mint}{lower sample time bound. Clusters must have sample before this time}

\item{maxt}{upper sample time bound. Clusters must have sample after this time}
}
\value{
data frame with log odds of new lineages or uk_lineage
}
\description{
This returns a data frame with lieange log odds broken down by epi week
}
