% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{cluster_sizes}
\alias{cluster_sizes}
\title{Compute cluster sizes only counting clusters with first sample occuring within given range and arrange by genotype}
\usage{
cluster_sizes(s, mint = 2020.2, maxt = 2020.35)
}
\arguments{
\item{s}{data frame with sequence_name(character), del_introduction(character), sample_time(numeric), genotype (wt or mutant)}

\item{mint}{lower sample time bound. Clusters must have sample before this time}

\item{maxt}{upper sample time bound. Clusters must have sample after this time}
}
\value{
data frame with cluster sizes, origin time, and genotype
}
\description{
Compute cluster sizes only counting clusters with first sample occuring within given range and arrange by genotype
}
