% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{clusterwise_logistic}
\alias{clusterwise_logistic}
\title{Estimate rate that clusters increase or decrease in frequency and transform into selection coefficients for different genotypes}
\usage{
clusterwise_logistic(
  s,
  mint = 2020.2,
  maxt = 2020.35,
  minClusterSize = 25,
  r = 73
)
}
\arguments{
\item{s}{data frame with sequence_name(character), del_introduction(character), sample_time(numeric), genotype (wt or mutant)}

\item{mint}{lower sample time bound. Clusters must have sample before this time}

\item{maxt}{upper sample time bound. Clusters must have sample after this time}

\item{minClusterSize}{exclude clusters with fewer samples than this}

\item{r}{specifies generation time (1/r) when transforming to selection coefficients. Default 73 corresponds to five days.}
}
\value{
list with matrix of selection coefficients for different clusters and a t.test for difference in 'mutant' selection coefficients
}
\description{
Estimate rate that clusters increase or decrease in frequency and transform into selection coefficients for different genotypes
}
