% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\name{compare_age_groups}
\alias{compare_age_groups}
\title{external wrapper for inner_compare_age_groups}
\usage{
compare_age_groups(
  targetnodes = NULL,
  scanner_env = readRDS("scanner-env-2021-03-03.rds"),
  path_to_data = "/cephfs/covid/bham/results/phylogenetics/latest/",
  include_pillar1 = F,
  min_date = NULL,
  max_date = NULL,
  threshold_growth = 0.5,
  fast_return = F
)
}
\description{
external wrapper for inner_compare_age_groups
}
