% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{compute_genotypes}
\alias{compute_genotypes}
\title{Compute AA genotype at specified postions in alignment}
\usage{
compute_genotypes(
  algn_fn = "alignments/cog_2020-11-11_alignment.fasta",
  positions = list(S614 = 23402:23404, S222 = 22226:22228, N220 = 28931:28933, ORF10 =
    29645:29647, Y453F = 22919:22921),
  ofn = NULL
)
}
\arguments{
\item{positions}{Named list of nt coordinates for given genotypes; each set of coordinates should span a codon}

\item{ofn}{optional path to csv for output}
}
\value{
data frame with genotypes
}
\description{
Compute AA genotype at specified postions in alignment
}
