% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\name{condense_clusters}
\alias{condense_clusters}
\title{wrapper for external calls to inner_condense_clusters}
\usage{
condense_clusters(scanner_env, threshold_growth = 0.5, candidate_nodes = NULL)
}
\description{
wrapper for external calls to inner_condense_clusters
}
