% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner_plots.R
\name{freq_plots}
\alias{freq_plots}
\title{Functions required to output scanner markdown figures
 
freq_plots}
\usage{
freq_plots(sdf = csd_select, kp_nodes = kp_nodes, growth_rank = growth_rank2)
}
\description{
Functions required to output scanner markdown figures
 
freq_plots
}
