% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\name{get_comparator_sample}
\alias{get_comparator_sample}
\title{matched by time and in proportion to adm2 prevalence}
\usage{
get_comparator_sample(u, scanner_env, nX = 5)
}
\description{
matched by time and in proportion to adm2 prevalence
}
