% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{hier_bayes_exponentialGrowth_frequency}
\alias{hier_bayes_exponentialGrowth_frequency}
\title{Hierarchical Bayesian inference of cluster-wise selection coefficients with a random effect corresponding to a dichotomous genotype}
\usage{
hier_bayes_exponentialGrowth_frequency(
  s,
  iterations = 8e+06,
  thin = 200,
  ncpu = 4,
  MINT = -Inf,
  MAXT = 2020.25,
  MU = 73,
  minClusterSize = 10
)
}
\arguments{
\item{s}{data frame that describes each sample, must have columns: del_introduction, sample_time, genotype}

\item{MINT}{minimum sample time for including cluster}

\item{MAXT}{exclude samples after this date}

\item{MU}{sets time scale of process and generation time. Generation time is 1/MU years}

\item{minClusterSize}{minimum size for cluster to be included in analysis}
}
\value{
Bayesian sampler output produced by BayesianTools::runMCMC
}
\description{
Hierarchical Bayesian inference of cluster-wise selection coefficients with a random effect corresponding to a dichotomous genotype
}
