% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{hier_bayes_pairwiseLogistic_frequency}
\alias{hier_bayes_pairwiseLogistic_frequency}
\title{too slow}
\usage{
hier_bayes_pairwiseLogistic_frequency(
  s,
  d,
  iterations = 8e+06,
  thin = 200,
  ncpu = 4,
  r = 73,
  minclustsize = 10
)
}
\description{
too slow
}
