% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\name{inner_compare_age_groups}
\alias{inner_compare_age_groups}
\title{internal function calls only, assuming amd is already in the environment}
\usage{
inner_compare_age_groups(u = 406318, fast_return = F)
}
\description{
internal function calls only, assuming amd is already in the environment
}
