% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\name{inner_condense_clusters}
\alias{inner_condense_clusters}
\title{internal function calls only}
\usage{
inner_condense_clusters(Y, threshold_growth = 0.5, candidate_nodes = NULL)
}
\description{
internal function calls only
}
