% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{matched_sample}
\alias{matched_sample}
\title{Matched sample by time (week) and UTLA}
\usage{
matched_sample(
  csids,
  lineage = NULL,
  not_lineage = "B.1.1.7",
  nreps = 1,
  n_multiplier = 1,
  deduplicate = FALSE
)
}
\arguments{
\item{csids}{central_sample_id to be matched}

\item{lineage}{If not NULL, sample will only include this lineage}

\item{not_lineage}{The sample will *not* include this lineage}

\item{nreps}{replicates}

\item{n_multiplier}{Optionally can draw multiple (integer) matches per element in csids}

\item{deduplicate}{not implemented}
}
\description{
Matched sample by time (week) and UTLA
}
