% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_mutations.R
\name{recurrent_mutations}
\alias{recurrent_mutations}
\title{Identify recurrent mutations in a tree and quantify offspring at nodes with mixed descendents}
\usage{
recurrent_mutations(tr1, tipstate, sts = NULL)
}
\arguments{
\item{tr1}{ape phylo, rooted, may have polytomies}

\item{tipstate}{character vector with names corresonding to tip labels; gives genotype}

\item{sts}{optional vector of sample times named according to tip label; gives approx age of mixed nodes}
}
\value{
Data frame with id of 'mixed' nodes that have offspring of multiple genotypes and the number of offspring of each genotype
}
\description{
Identify recurrent mutations in a tree and quantify offspring at nodes with mixed descendents
}
