% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_lineage}
\alias{sample_lineage}
\title{Produce a random sample of given global lineage and sample weights}
\usage{
sample_lineage(
  lineage = "B.1.1.7",
  weightsfn = "/cephfs/covid/bham/climb-covid19-volze/b0-weightsdf-2021-01-04.csv",
  mindate = as.Date("2020-10-15"),
  maxdate = as.Date(Sys.Date() - 12),
  nreps = 100,
  ns = c(250, 500, 750, 1000),
  prop_stratified = 0.25,
  deduplicate = TRUE
)
}
\arguments{
\item{lineage}{Global lineage}

\item{weightsfn}{output produced by coverage_weights}

\item{mindate}{minimum date}

\item{maxdate}{maximum date; note it takes about 12 days for P2 sampling to stabilize}

\item{nreps}{replicates}

\item{ns}{sample sizes to use}

\item{prop_stratified}{proportion of sample to reserve for stratified sampling}

\item{deduplicate}{if TRUE, duplicate samples will be removed}
}
\description{
Sampling is repeated multiple times and over multiple sample sizes. Can confine sampling to a particular time frame. A proportion of samples can be stratified through time rather than being drawn completely at random (weighted smapling is still used within each week)
}
