% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\name{scanner}
\alias{scanner}
\title{Compute scannter statistics for nodes in tree with >x descendants}
\usage{
scanner(
  treenexfn = NULL,
  min_descendants = 30,
  max_descendants = 20000,
  min_date = NULL,
  max_date = NULL,
  ncpu = 8,
  path_to_data = "/cephfs/covid/bham/results/phylogenetics/latest/"
)
}
\arguments{
\item{treenexfn}{File name of tree with metadata in node labels}

\item{min_descendants}{Clade must have at least this many tips}

\item{max_descendants}{Clade can have at most this many tips}

\item{min_date}{Only include samples after this data}

\item{max_date}{Only include samples before and including this date}

\item{ncpu}{number cpu for multicore ops}

\item{path_to_data}{Path to data with COG tree and COG tree metadata}
}
\description{
Computes a logistic growth rate and a statistic for outlying values in a molecular clock (root to tip regression).
Only counts pillar 1 data; comparison sample is matched by time and adm2.
}
