% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner2.R
\name{scanner2}
\alias{scanner2}
\title{Compute scanner statistics for nodes in tree.}
\usage{
scanner2(
  tre,
  amd,
  min_descendants = 30,
  max_descendants = 20000,
  min_cluster_age_yrs = 1.5/12,
  min_date = NULL,
  max_date = NULL,
  min_blen = 0,
  ncpu = 8,
  output_dir = ".",
  include_pillar1 = TRUE,
  country = NULL,
  num_ancestor_comparison = 500,
  factor_geo_comparison = 5,
  Tg = 6.5/365,
  report_freq = 50,
  compute_treestructure = FALSE,
  lsd_args = list(givenRate = 0.00075, seqLen = 29000, variance = 0, verbose = TRUE),
  ts_args = list(minCladeSize = 50, level = 0.001),
  root_on_tip = "Wuhan/WH04/2020",
  root_on_tip_sample_time = 2020
)
}
\arguments{
\item{tre}{A rooted phylogeny in ape::phylo or treeio::treedata form. The latter may include internal node labels and be read by '.read.beast'.}

\item{amd}{A data frame containing required metadata for each tip in tree: sequence_name, sample_date, region. Optional metadata includes: pillar_2, sample_time(numeric), country,  lineage.}

\item{min_descendants}{Clade must have at least this many tips}

\item{max_descendants}{Clade can have at most this many tips}

\item{min_cluster_age_yrs}{Only include clades that have sample tips that span at least this value}

\item{min_date}{Only include samples after this data}

\item{max_date}{Only include samples before and including this date}

\item{min_blen}{Only compute statistics for nodes descended from branches of at least this length}

\item{ncpu}{number cpu for multicore ops}

\item{output_dir}{Path to directory where results will be saved}

\item{include_pillar1}{if TRUE (default TRUE), will include Pillar 1 samples when computing stats}

\item{country}{Optional, only include data from this country}

\item{num_ancestor_comparison}{When finding comparison sample for molecular clock stat, make sure sister clade has at least this many tips}

\item{factor_geo_comparison}{When finding comparison sample based on geography, make sure sample has this factor times the number within clade of interest}

\item{Tg}{Approximate generation time in years}

\item{report_freq}{Print progress for every n'th node}

\item{compute_treestructure}{If TRUE, will compute treestructure clusters and ensure other tree stats are computed for these clusters}
}
\description{
Takes standard inputs in the form of a rooted phylogeny and data frame with required metadata (see below). 
Computes a logistic growth rate and a statistic for outlying values in a molecular clock (root to tip regression).
Comparison sample is matched by time and region.
}
\details{
add optional 'sample_weight' to metadata, use in regression models
}
