% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\name{scanner_logistic_growth_rate}
\alias{scanner_logistic_growth_rate}
\title{Compute growth rate statistic for nodes in tree with >x descendants}
\usage{
scanner_logistic_growth_rate(
  min_descendants = 50,
  max_descendants = 20000,
  min_date = NULL,
  max_date = NULL,
  ncpu = 8,
  path_to_data = "/cephfs/covid/bham/results/phylogenetics/latest/"
)
}
\arguments{
\item{min_descendants}{Clade must have at least this many tips}

\item{max_descendants}{Clade can have at most this many tips}

\item{min_date}{Only include samples after this data}

\item{max_date}{Only include samples before and including this date}

\item{ncpu}{number cpu for multicore ops}

\item{path_to_data}{Path to data with COG tree and COG tree metadata}
}
\description{
Only counts pillar 1 data; comparison sample is matched by time and adm2
}
