% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.R
\name{sim_cluster_growth}
\alias{sim_cluster_growth}
\title{Simulate cluster trajectories}
\usage{
sim_cluster_growth(
  tstart,
  tfin,
  R,
  Rsd = 0.2,
  non_extinction = 0,
  ntries = 10,
  seir_gen = NULL,
  ...
)
}
\arguments{
\item{tstart}{Time of origin}

\item{tfin}{Simulate until}

\item{R}{Expected reproduction number}

\item{Rsd}{std dev of R}

\item{non_extinction}{Minimum cumulative infections required to return trajectory}

\item{...}{additional arguments are passed to seir model generator (e.g. for changing initial conditions)}
}
\value{
Matrix with simulation output
}
\description{
SDE model for simple exponential growth, stores cumulative and active infections
}
