% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.R
\name{sim_importation_time}
\alias{sim_importation_time}
\title{Simulate time and number of importation events for clusters of two lineages}
\usage{
sim_importation_time(E_imports0, E_imports1, mu, sigma)
}
\arguments{
\item{E_imports0}{Expected number of imports for variant 0 which initiate cluster}

\item{mu}{Mean time of import}

\item{sigma}{std dev of importation time}
}
\value{
$variant and $ancestral contain times of import
}
\description{
Simulate time and number of importation events for clusters of two lineages
}
