% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.R
\name{sim_replicate}
\alias{sim_replicate}
\title{Simulate entire pipeline: importation, growth, sampling}
\usage{
sim_replicate(
  E_imports0 = 650,
  E_imports1 = 650,
  mu = lubridate::decimal_date(as.Date("2021-04-18")),
  sigma = 12.9/365,
  tfin = lubridate::decimal_date(as.Date("2021-04-18")) + 1/12,
  Rvariant = 1.5 * 1,
  Rancestral = 0.9,
  Rsd = 0.2,
  rho0 = 0.9,
  rho1 = 0.66
)
}
\description{
Simulate entire pipeline: importation, growth, sampling
}
