% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.R
\name{sim_replicate2}
\alias{sim_replicate2}
\title{Uses hier bayes model for identifying growth}
\usage{
sim_replicate2(
  MU = lubridate::decimal_date(as.Date("2021-04-18")),
  TFIN,
  RHO0,
  k = 25,
  bs_replicates = 100,
  ncpu = 1,
  ...
)
}
\description{
Uses hier bayes model for identifying growth
}
