% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.R
\name{sim_sampling_cluster}
\alias{sim_sampling_cluster}
\title{Simulate sampling times given cluster trajectory and sample rate}
\usage{
sim_sampling_cluster(X, rho0 = 0.1, rho1 = 0.25)
}
\arguments{
\item{X}{output of sim_cluster_growth}

\item{rho0}{probability that case is sampled}

\item{rho1}{probability that case is diagnosed}
}
\value{
vector of sample times
}
\description{
Note sampling rate decays to zero from tfin - 12 days to tfin
}
