% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner_plots.R
\name{variable_frequency_day_report}
\alias{variable_frequency_day_report}
\title{GAM outputs}
\usage{
variable_frequency_day_report(
  s,
  variable = "type",
  value = "clade",
  mint = -Inf,
  maxt = Inf,
  detailed = TRUE,
  form = y ~ s(sample_time, bs = "gp", k = 5)
)
}
\description{
GAM outputs
}
