% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{variable_frequency_epiweek}
\alias{variable_frequency_epiweek}
\title{Simple log odds freqency plot for given variable. Does not partition by cluster.}
\usage{
variable_frequency_epiweek(
  s,
  variable = "genotype",
  value = "mutant",
  mint = -Inf,
  maxt = Inf,
  detailed = TRUE,
  form = y ~ s(sample_time, bs = "gp", k = 5)
)
}
\arguments{
\item{s}{data frame which must contain sample_time and 'variable' for each sequence}

\item{variable}{Column of s which contains variable to split the data}

\item{value}{The value in conlumn 'variable' for which we will compute frequency}

\item{mint}{Minimum sample time to include}

\item{maxt}{Max time to include}

\item{form}{formula passed to mgcv::gam for estimating frequencies}
}
\value{
list with data, plots, and estimated frequencies
}
\description{
If the column 'weight' is in the input data frame, this will use the values to weight each observation. 
If computing for genotypes you will want to remove gaps and 'X' before passing to this function; also deduplicate by patient id.
}
