\name{variantAnalysis-package}
\alias{variantAnalysis-package}
\alias{variantAnalysis}
\docType{package}
\title{
\packageTitle{variantAnalysis}
}
\description{
\packageDescription{variantAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{variantAnalysis}
\packageIndices{variantAnalysis}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{variantAnalysis}

Maintainer: \packageMaintainer{variantAnalysis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
