% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{hier_bayes_exponentialGrowth_frequency}
\alias{hier_bayes_exponentialGrowth_frequency}
\title{Hierarchical Bayesian inference of cluster-wise selection coefficients with a random effect corresponding to a dichotomous genotype}
\usage{
hier_bayes_exponentialGrowth_frequency(
  s,
  d,
  iterations = 8e+06,
  thin = 200,
  ncpu = 4,
  MINT = -Inf,
  MAXT = 2020.25,
  MU = 73
)
}
\arguments{
\item{s}{data frame that describes each sample, must have columns: del_introduction, sample_time}

\item{d}{data frame that describes each cluster, must have columns: lineage (matching del_introduction), first_sample (time of first sample in cluster), genotype ( 'wt' or 'mutant' )}

\item{MINT}{minimum sample time for including cluster}

\item{MAXT}{exclude samples after this date}

\item{MU}{sets time scale of process and generation time. Generation time is 1/MU years}
}
\value{
Bayesian sampler output produced by BayesianTools::runMCMC
}
\description{
Hierarchical Bayesian inference of cluster-wise selection coefficients with a random effect corresponding to a dichotomous genotype
}
