% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{matched_sample2}
\alias{matched_sample2}
\title{Matched sample by time (week) and UTLA for each sampled tree output from sample_lineage function}
\usage{
matched_sample2(
  sample_lineage_output,
  lineage = "B.1.177",
  not_lineage = "B.1.1.7",
  n_multiplier = 1,
  deduplicate = FALSE
)
}
\arguments{
\item{sample_lineage_output}{output from sample_lineage}

\item{lineage}{If not NULL, sample will only include this lineage}

\item{not_lineage}{The sample will *not* include this lineage}

\item{n_multiplier}{Optionally can draw multiple (integer) matches per element in csids}

\item{deduplicate}{not implemented}
}
\description{
Matched sample by time (week) and UTLA for each sampled tree output from sample_lineage function
}
