% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner2.R
\name{scanner2}
\alias{scanner2}
\title{Compute scanner statistics for nodes in tree.}
\usage{
scanner2(
  tre,
  amd,
  min_descendants = 30,
  max_descendants = 20000,
  min_date = NULL,
  max_date = NULL,
  ncpu = 8,
  output_dir = ".",
  include_pillar1 = FALSE,
  country = NULL,
  num_ancestor_comparison = 500,
  factor_geo_comparison = 5,
  Tg = 6.5/365
)
}
\arguments{
\item{tre}{A rooted phylogeny in ape::phylo or treeio::treedata form. The latter may include internal node labels and be read by '.read.beast'.}

\item{amd}{A data frame containing required metadata for each tip in tree: sequence_name, sample_date, region. Optional metadata includes: pillar_2, sample_time(numeric), country.}

\item{min_descendants}{Clade must have at least this many tips}

\item{max_descendants}{Clade can have at most this many tips}

\item{min_date}{Only include samples after this data}

\item{max_date}{Only include samples before and including this date}

\item{ncpu}{number cpu for multicore ops}

\item{output_dir}{Path to directory where results will be saved}

\item{include_pillar1}{if TRUE (default FALSE), will include Pillar 1 samples when computing stats}

\item{country}{Optional, only include data from this country}

\item{num_ancestor_comparison}{When finding comparison sample for molecular clock stat, make sure sister clade has at least this many tips}

\item{factor_geo_comparison}{When finding comparison sample based on geography, make sure sample has this factor times the number within clade of interest}

\item{generation_time_scale}{Approximate generation time in years}
}
\description{
Takes standard inputs in the form of a rooted phylogeny and data frame with required metadata (see below). 
Computes a logistic growth rate and a statistic for outlying values in a molecular clock (root to tip regression).
Comparison sample is matched by time and region.
}
\details{
TODO: include patientid in metadata & deduplicate 
add optional 'sample_weight' to metadata, use in regression models
}
