% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power1.R
\name{sim_replicate}
\alias{sim_replicate}
\title{Simulate entire pipeline: importation, growth, sampling}
\usage{
sim_replicate(
  E_imports0,
  E_imports1,
  mu = 2021.299,
  sigma = 1/12/4,
  tfin = 2021.299 + 1.5/12,
  Rvariant = 2,
  Rancestral = 1.5,
  rho0 = 0.1,
  rho1 = 0.5
)
}
\description{
Simulate entire pipeline: importation, growth, sampling
}
