% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency0.R
\name{variable_frequency_epiweek}
\alias{variable_frequency_epiweek}
\title{Simple log odds freqency plot for given variable. Does not partition by cluster.}
\usage{
variable_frequency_epiweek(
  s,
  variable = "genotype",
  value = "mutant",
  mint = -Inf,
  maxt = Inf,
  detailed = FALSE
)
}
\value{
data frame with frequency data. If detailed==TRUE, also returns data and a plot with loess regression line
}
\description{
If computing for genotypes you will want to remove gaps and 'X' before passing to this function; also deduplicate by patient id
}
